Move menu_level_bed_corners to its own file
